Support for filteredScopes, multi project/cluster access control
This supports the latest prom-keycloak-proxy features for regex
filters in prom-keycloak-proxy for access control to multiple
individual cluster and project permissions at the same time.
Adding this to one of the dashboards, but this will need to be
applied to the rest of the custom dashboards in HubSearchPage.htm,
ClusterSearchPage.htm, ProjectSearchPage.htm, and possibly more.