Fix duplicate series issue for gpu hours dashboard
The problem causing this bug was that the query within the queryGpuBillingHours in site.js was joining kube_pod_resource_request with kube_node_labels (we join with kube_node_labels in order to get the GPU model name) but kube_node_labels can have multiple time series for the same node. The fix is to use max by so the query only returns one series.

Signed-off-by: IsaiahStapleton <istaplet@redhat.com>